+++ src/lib/libast/string/fmtelapsed.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 /*
@@ -25,15 +26,15 @@
 #include <ast.h>
 
 char*
-fmtelapsed(unsigned long u, int n)
+fmtelapsed(unsigned long u, unsigned long n)
 {
 	unsigned long	t;
 	char*		buf;
-	int		z;
+	size_t		z;
 
-	if (u == 0L)
+	if (u == 0UL)
 		return "0";
-	if (u == ~0L)
+	if (u == ~0UL)
 		return "%";
 	buf = fmtbuf(z = 8);
 	t = u / n;
